fix: address review findings on settings DI naming and TabHistory reactivity
Move settings DI wiring out of PersistenceModule.kt into a dedicated
di/SettingsModule.kt (SettingsProvideModule + SettingsBindModule), matching
GetIcon's naming exactly per the fleet-wide settings-pattern mandate.
PersistenceModule.kt goes back to owning only Room/DAO provisioning.

Add UserSettings.errorLimitFlow and switch TabHistory back to a continuous
Flow observation (matching the original DataStore-era shape) instead of the
onResume()-only refresh from the previous commit, which silently dropped
reactivity while the tab is visible-but-unfocused (e.g. split-screen next to
Settings) since onResume() doesn't fire in that state.

Also drops the now-unused de.lemke.commonutils.data.sanitized import from
UserSettings.kt.
