+++ app/src/main/java/de/lemke/sudoku/PersistenceModule.kt
@@ -18,45 +18,21 @@ package de.lemke.sudoku
 
 import android.app.Application
 import android.content.Context
-import androidx.preference.PreferenceManager
 import androidx.room.Room
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import de.lemke.commonutils.data.SettingsRepository
-import de.lemke.sudoku.data.UserSettings
 import de.lemke.sudoku.data.database.AppDatabase
 import de.lemke.sudoku.data.database.MIGRATION_1_2
 import de.lemke.sudoku.data.database.SudokuDao
-import javax.inject.Qualifier
 import javax.inject.Singleton
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-
-@Retention(AnnotationRetention.RUNTIME)
-@Qualifier
-annotation class ApplicationScope
 
 @Module
 @InstallIn(SingletonComponent::class)
 object
 PersistenceModule : Application() {
-    @Provides
-    @Singleton
-    @ApplicationScope
-    fun providesApplicationScope(): CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
-
-    @Provides
-    @Singleton
-    fun provideUserSettings(
-        @ApplicationContext context: Context,
-        @ApplicationScope scope: CoroutineScope,
-    ): UserSettings = UserSettings(PreferenceManager.getDefaultSharedPreferences(context), scope)
-
     @Provides
     @Singleton
     fun provideAppDatabase(
@@ -70,14 +46,4 @@ PersistenceModule : Application() {
 
     @Provides
     fun provideSudokuDao(database: AppDatabase): SudokuDao = database.sudokuDao()
-}
-
-// common-utils' own activities (CommonUtilsAboutActivity, CommonUtilsSettingsActivity.SettingsFragment) inject
-// SettingsRepository directly via Hilt, so it must be bound in the graph regardless of app-specific settings access
-// (mirrors GetIcon's SettingsBindModule).
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class SettingsBindModule {
-    @Binds
-    abstract fun bindSettingsRepository(userSettings: UserSettings): SettingsRepository
 }
