+++ app/src/main/java/de/lemke/sudoku/di/SettingsModule.kt
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2022-2026 Leonard Lemke
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.lemke.sudoku.di
+
+import android.content.Context
+import androidx.preference.PreferenceManager
+import dagger.Binds
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import de.lemke.commonutils.data.SettingsRepository
+import de.lemke.sudoku.data.UserSettings
+import javax.inject.Qualifier
+import javax.inject.Singleton
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+
+@Retention(AnnotationRetention.RUNTIME)
+@Qualifier
+annotation class ApplicationScope
+
+@Module
+@InstallIn(SingletonComponent::class)
+object SettingsProvideModule {
+    @Provides
+    @Singleton
+    @ApplicationScope
+    fun providesApplicationScope(): CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
+
+    @Provides
+    @Singleton
+    fun provideUserSettings(
+        @ApplicationContext c: Context,
+        @ApplicationScope scope: CoroutineScope,
+    ): UserSettings = UserSettings(PreferenceManager.getDefaultSharedPreferences(c), scope)
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class SettingsBindModule {
+    @Binds
+    abstract fun bindSettingsRepository(u: UserSettings): SettingsRepository
+}
